Move decorators from utils module to decorators module.
